#26 update : main navigation 처리 분리
